tweak image permissions so that users with only add_image permission can only edit/delete their own images

+++ django-verdant/verdantimages/views/images.py
@@ -1,7 +1,8 @@
 from django.shortcuts import render, redirect, get_object_or_404
 from django.contrib import messages
 from django.core.paginator import Paginator, EmptyPage, PageNotAnInteger
-from django.contrib.auth.decorators import permission_required
+from django.contrib.auth.decorators import permission_required, login_required
+from django.core.exceptions import PermissionDenied
 
 from verdantimages.models import get_image_model
 from verdantimages.forms import get_image_form
@@ -24,8 +25,14 @@ def index(request):
             images = Image.search(q, results_per_page=20, page=p)
         else:
             images = Image.objects.order_by('-created_at')
+            if not request.user.has_perm('verdantimages.change_image'):
+                # restrict to the user's own images
+                images = images.filter(uploaded_by_user=request.user)
     else:
         images = Image.objects.order_by('-created_at')
+        if not request.user.has_perm('verdantimages.change_image'):
+            # restrict to the user's own images
+            images = images.filter(uploaded_by_user=request.user)
         form = SearchForm()
 
     if not is_searching:
@@ -54,12 +61,15 @@ def index(request):
         })
 
 
-@permission_required('verdantimages.change_image')
+@login_required  # more specific permission tests are applied within the view
 def edit(request, image_id):
     Image = get_image_model()
     ImageForm = get_image_form()
 
     image = get_object_or_404(Image, id=image_id)
+
+    if not image.is_editable_by_user(request.user):
+        raise PermissionDenied
 
     if request.POST:
         original_file = image.file
@@ -85,9 +95,12 @@ def edit(request, image_id):
     })
 
 
-@permission_required('verdantimages.delete_image')
+@login_required  # more specific permission tests are applied within the view
 def delete(request, image_id):
     image = get_object_or_404(get_image_model(), id=image_id)
+
+    if not image.is_editable_by_user(request.user):
+        raise PermissionDenied
 
     if request.POST:
         image.delete()
@@ -102,11 +115,13 @@ def delete(request, image_id):
 @permission_required('verdantimages.add_image')
 def add(request):
     ImageForm = get_image_form()
+    ImageModel = get_image_model()
 
     if request.POST:
-        form = ImageForm(request.POST, request.FILES)
+        image = ImageModel(uploaded_by_user=request.user)
+        form = ImageForm(request.POST, request.FILES, instance=image)
         if form.is_valid():
-            image = form.save()
+            form.save()
             messages.success(request, "Image '%s' added." % image.title)
             return redirect('verdantimages_index')
         else:
